requirements.txt
fastapi
uvicorn[standard]
python-dotenv
langchain
langchain-mistralai
langchain-chroma
chromadb
faster-whisper
react-markdown
pytest
transformers
Pillow
torch

